Move tests into appropriate package
