+++ src/test/java/seedu/address/model/displayable/seller/SellerTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.displayable;
+package seedu.address.model.displayable.seller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
